+++ src/com/rian/difficultycalculator/beatmap/timings/BreakPeriod.java
@@ -7,18 +7,18 @@ public final class BreakPeriod {
     /**
      * The time at which this break period starts, in milliseconds.
      */
-    public final int startTime;
+    public final double startTime;
 
     /**
      * The time at which this break period ends, in milliseconds.
      */
-    public final int endTime;
+    public final double endTime;
 
     /**
      * @param startTime The time at which this break period starts, in milliseconds.
      * @param endTime The time at which this break period ends, in milliseconds.
      */
-    public BreakPeriod(int startTime, int endTime) {
+    public BreakPeriod(double startTime, double endTime) {
         this.startTime = startTime;
         this.endTime = endTime;
     }
@@ -26,7 +26,7 @@ public final class BreakPeriod {
     /**
      * Gets the duration of this break period.
      */
-    public int getDuration() {
+    public double getDuration() {
         return endTime - startTime;
     }
 }
